fixes mask comparison to compare up to the mask length
pkt_match was strictly comparing the length of the packet argument
with the size of the mask. It is desirable to compare only the
prefix given by the mask.size.

Also fixed the printing to match how we print packet output using
hexdump -- it captures the print output in a buffer that is then
returned.