Better --test-params handling
Log an error message when the string passed to --test-params cannot be
evaluate by Python's exec method.

Also includes some minor improvement to logging. In particular, some scapy
error logs are suppressed and we now log errors to stdout.

We test the new --test-params code with some Python (python3) unit
tests. Unfortunately ptf is not very unit test friendly at the moment so we
actually start a ptf process with subprocess and inspect stdout /
stderr. Ideally, the ptf script would be change in the future so that ptf can be
run as a function and methods can be independently tested.