ptf: testutils: add parameter vxlan_reserved0 to simple_vxlan_packet
There is an issue with VxLAN API when you try to send VxLAN packet
and compare with expected VxLAN pkt received on another port for example.
verify_packet returns assert, even if you send VxLAN packet and expect
the same VxLAN packet (with the same fields).

The problem is that vxlan_reserved0 value of sent packet and received packets
is different (0 vs None). So, to fix this, make sure this filed is zero.

Signed-off-by: Oleh Yakymenko <olehx.yakymenko@intel.com>
Signed-off-by: Volodymyr Mytnyk <volodymyrx.mytnyk@intel.com>